Parametrized the algorithm names in the testing. Made the search for algorithm files a little more robust.
